+++ src/main/java/org/opentripplanner/framework/application/OTPFeature.java
@@ -105,8 +105,7 @@ public enum OTPFeature {
   SandboxAPIMapboxVectorTilesApi(false, true, "Enable Mapbox vector tiles API."),
   SandboxAPIParkAndRideApi(false, true, "Enable park-and-ride endpoint."),
   SandboxAPITravelTime(false, true, "Enable the isochrone/travel time surface API."),
-  TransferAnalyzer(false, true, "Analyze transfers during graph build."),
-  VehicleToStopHeuristics(false, true, "Enable improved heuristic for park-and-ride queries.");
+  TransferAnalyzer(false, true, "Analyze transfers during graph build.");
 
   private static final Object TEST_LOCK = new Object();