+++ src/main/java/org/opentripplanner/transit/service/DefaultTransitService.java
@@ -41,6 +41,7 @@ import org.opentripplanner.transit.model.network.TripPattern;
 import org.opentripplanner.transit.model.organization.Agency;
 import org.opentripplanner.transit.model.organization.Operator;
 import org.opentripplanner.transit.model.site.AreaStop;
+import org.opentripplanner.transit.model.site.GroupStop;
 import org.opentripplanner.transit.model.site.MultiModalStation;
 import org.opentripplanner.transit.model.site.RegularStop;
 import org.opentripplanner.transit.model.site.Station;
@@ -224,6 +225,12 @@ public class DefaultTransitService implements TransitEditorService {
   public Collection<RegularStop> listRegularStops() {
     OTPRequestTimeoutException.checkForTimeout();
     return transitModel.getStopModel().listRegularStops();
+  }
+
+  @Override
+  public Collection<GroupStop> listGroupStops() {
+    OTPRequestTimeoutException.checkForTimeout();
+    return transitModel.getStopModel().listGroupStops();
   }
 
   @Override
@@ -543,7 +550,7 @@ public class DefaultTransitService implements TransitEditorService {
   }
 
   @Override
-  public Collection<RegularStop> findRegularStop(Envelope envelope) {
+  public Collection<RegularStop> findRegularStops(Envelope envelope) {
     OTPRequestTimeoutException.checkForTimeout();
     return transitModel.getStopModel().findRegularStops(envelope);
   }
@@ -551,7 +558,7 @@ public class DefaultTransitService implements TransitEditorService {
   @Override
   public Collection<AreaStop> findAreaStops(Envelope envelope) {
     OTPRequestTimeoutException.checkForTimeout();
-    return transitModel.getStopModel().queryLocationIndex(envelope);
+    return transitModel.getStopModel().findAreaStops(envelope);
   }
 
   @Override