+++ src/test/java/org/opentripplanner/graph_builder/module/osm/TriangleInequalityTest.java
@@ -182,6 +182,7 @@ public class TriangleInequalityTest {
         .withStreet(street -> street.withTurnReluctance(1.0))
         .withCar(car -> car.withSpeed(1.0).withReluctance(1.0))
         .withBike(bike -> bike.withSpeed(1.0).withReluctance(1.0))
+        .withScooter(scooter -> scooter.withSpeed(1.0).withReluctance(1.0))
     );
 
     if (modes != null) {