+++ src/test/java/org/opentripplanner/routing/algorithm/TurnCostTest.java
@@ -94,6 +94,7 @@ public class TurnCostTest {
       preferences
         .withCar(it -> it.withSpeed(1.0).withReluctance(1.0))
         .withBike(bike -> bike.withSpeed(1.0).withReluctance(1.0))
+        .withScooter(scooter -> scooter.withSpeed(1.0).withReluctance(1.0))
         .withWalk(walk -> walk.withSpeed(1.0).withStairsReluctance(1.0).withReluctance(1.0))
         .withStreet(it -> it.withTurnReluctance(1.0))
     );